+++ src/deprecated.jl
@@ -77,7 +77,7 @@ Broadcast `f` on the input.
   - Broadcasted Activation `f.(x)`
   - Empty `NamedTuple()`
 
-!!! warn
+!!! warning
     
     This layer is deprecated and will be removed in v0.5. Use [`WrappedFunction`](@ref) with
     manual broadcasting