+++ src/layers/conv.jl
@@ -1,20 +1,19 @@
 @doc doc"""
     Conv(k::NTuple{N,Integer}, (in_chs => out_chs)::Pair{<:Integer,<:Integer},
-         activation=identity; init_weight=glorot_uniform, stride=1, pad=0,
-         dilation=1, groups=1, bias=true)
+         activation=identity; init_weight=glorot_uniform, init_bias=zeros32, stride=1,
+         pad=0, dilation=1, groups=1, use_bias=true)
 
 Standard convolutional layer.
 
 Image data should be stored in WHCN order (width, height, channels, batch). In other words,
-a `100 × 100` RGB image would be a `100 × 100 × 3 × 1` array, and a batch of 50 would be a
-`100 × 100 × 3 × 50` array. This has `N = 2` spatial dimensions, and needs a kernel size
+a `100 x 100` RGB image would be a `100 x 100 x 3 x 1` array, and a batch of 50 would be a
+`100 x 100 x 3 x 50` array. This has `N = 2` spatial dimensions, and needs a kernel size
 like `(5, 5)`, a 2-tuple of integers. To take convolutions along `N` feature dimensions,
 this layer expects as input an array with `ndims(x) == N + 2`, where
 `size(x, N + 1) == in_chs` is the number of input channels, and `size(x, ndims(x))` is the
 number of observations in a batch.
 
 !!! note
-    
     Frameworks like [`Pytorch`](https://pytorch.org/docs/stable/generated/torch.nn.Conv2d.html#torch.nn.Conv2d)
     perform cross-correlation in their convolution layers
 
@@ -29,6 +28,7 @@ number of observations in a batch.
 ## Keyword Arguments
 
   - `init_weight`: Controls the initialization of the weight parameter
+  - `init_bias`: Controls the initialization of the bias parameter
 
   - `stride`: Should each be either single integer, or a tuple with `N` integers
   - `dilation`: Should each be either single integer, or a tuple with `N` integers
@@ -46,8 +46,7 @@ number of observations in a batch.
   - `groups`: Expected to be an `Int`. It specifies the number of groups to divide a
               convolution into (set `groups = in_chs` for Depthwise Convolutions). `in_chs`
               and `out_chs` must be divisible by `groups`.
-  - `bias`: The initial bias vector is all zero by default. Trainable bias can be disabled
-            entirely by setting this to `false`.
+  - `use_bias`: Trainable bias can be disabled entirely by setting this to `false`.
 
 ## Inputs
 
@@ -69,7 +68,7 @@ O_i = floor\left(\frac{I_i + pad[i] + pad[(i + N) \% length(pad)] - dilation[i] 
   - `weight`: Convolution kernel
   - `bias`: Bias (present if `bias=true`)
 """
-struct Conv{N, bias, M, F1, F2} <: AbstractExplicitLayer
+struct Conv{N, use_bias, M, F1, F2, F3} <: AbstractExplicitLayer
     activation::F1
     in_chs::Int
     out_chs::Int
@@ -79,55 +78,56 @@ struct Conv{N, bias, M, F1, F2} <: AbstractExplicitLayer
     dilation::NTuple{N, Int}
     groups::Int
     init_weight::F2
+    init_bias::F3
 end
 
-function Conv(k::NTuple{N, Integer},
-              ch::Pair{<:Integer, <:Integer},
-              activation=identity;
-              init_weight=glorot_uniform,
-              stride=1,
-              pad=0,
-              dilation=1,
-              groups=1,
-              bias=true) where {N}
-    stride = expand(Val(N), stride)
-    dilation = expand(Val(N), dilation)
+function Conv(k::NTuple{N, Integer}, ch::Pair{<:Integer, <:Integer}, activation=identity;
+              init_weight=glorot_uniform, init_bias=zeros32, stride=1, pad=0, dilation=1,
+              groups=1, use_bias::Bool=true, bias::Union{Missing, Bool}=missing) where {N}
+
+    # Deprecated Functionality (Remove in v0.5)
+    if !ismissing(bias)
+        Base.depwarn("`bias` argument to `Conv` has been deprecated and will be removed" *
+                     " in v0.5. Use `use_bias` kwarg instead.", :Conv)
+        if !use_bias
+            throw(ArgumentError("Both `bias` and `use_bias` are set. Please only use " *
+                                "the `use_bias` keyword argument."))
+        end
+        use_bias = bias
+    end
+
+    stride = _expand(Val(N), stride)
+    dilation = _expand(Val(N), dilation)
     pad = calc_padding(Conv, pad, k, dilation, stride)
     activation = NNlib.fast_act(activation)
-    return Conv{N, bias, length(pad), typeof(activation), typeof(init_weight)}(activation,
-                                                                               first(ch),
-                                                                               last(ch), k,
-                                                                               stride, pad,
-                                                                               dilation,
-                                                                               groups,
-                                                                               init_weight)
+
+    return Conv{N, use_bias, length(pad), typeof(activation), typeof(init_weight),
+                typeof(init_bias)}(activation, first(ch), last(ch), k, stride, pad,
+                                   dilation, groups, init_weight, init_bias)
 end
 
-function initialparameters(rng::AbstractRNG, c::Conv{N, bias}) where {N, bias}
-    weight = convfilter(rng, c.kernel_size, c.in_chs => c.out_chs; init=c.init_weight,
-                        groups=c.groups)
-    if bias
-        return (weight=weight,
-                bias=zeros(eltype(weight), ntuple(_ -> 1, N)..., c.out_chs, 1))
+function initialparameters(rng::AbstractRNG, c::Conv{N, use_bias}) where {N, use_bias}
+    weight = _convfilter(rng, c.kernel_size, c.in_chs => c.out_chs; init=c.init_weight,
+                         groups=c.groups)
+    if use_bias
+        return (weight=weight, bias=c.init_bias(rng, ntuple(_ -> 1, N)..., c.out_chs, 1))
     else
         return (weight=weight,)
     end
 end
 
-function parameterlength(c::Conv{N, bias}) where {N, bias}
-    return prod(c.kernel_size) * c.in_chs * c.out_chs ÷ c.groups + (bias ? c.out_chs : 0)
+function parameterlength(c::Conv{N, use_bias}) where {N, use_bias}
+    return prod(c.kernel_size) * c.in_chs * c.out_chs ÷ c.groups +
+           (use_bias ? c.out_chs : 0)
 end
 
-@inline function (c::Conv{N, false})(x::AbstractArray,
-                                     ps,
-                                     st::NamedTuple) where {N}
+@inline function (c::Conv{N, false})(x::AbstractArray, ps, st::NamedTuple) where {N}
     cdims = DenseConvDims(x, ps.weight; stride=c.stride, padding=c.pad, dilation=c.dilation,
                           groups=c.groups)
     return applyactivation(c.activation, conv_wrapper(x, ps.weight, cdims)), st
 end
 
-@inline function (c::Conv{N, true})(x::AbstractArray, ps,
-                                    st::NamedTuple) where {N}
+@inline function (c::Conv{N, true})(x::AbstractArray, ps, st::NamedTuple) where {N}
     cdims = DenseConvDims(x, ps.weight; stride=c.stride, padding=c.pad, dilation=c.dilation,
                           groups=c.groups)
     return applyactivation(c.activation,
@@ -201,7 +201,7 @@ struct MaxPool{N, M} <: AbstractExplicitLayer
 end
 
 function MaxPool(k::NTuple{N, Integer}; pad=0, stride=k) where {N}
-    stride = expand(Val(N), stride)
+    stride = _expand(Val(N), stride)
     pad = calc_padding(MaxPool, pad, k, 1, stride)
     return MaxPool{N, length(pad)}(k, pad, stride)
 end
@@ -268,7 +268,7 @@ struct MeanPool{N, M} <: AbstractExplicitLayer
 end
 
 function MeanPool(k::NTuple{N, Integer}; pad=0, stride=k) where {N}
-    stride = expand(Val(N), stride)
+    stride = _expand(Val(N), stride)
     pad = calc_padding(MeanPool, pad, k, 1, stride)
     return MeanPool{N, length(pad)}(k, pad, stride)
 end
@@ -297,7 +297,7 @@ Upsampling Layer.
 
   - `mode`: Set to `:nearest`, `:linear`, `:bilinear` or `:trilinear`
 
-One of two keywords must be given:
+Exactly one of two keywords must be specified:
 
   - If `scale` is a number, this applies to all but the last two dimensions (channel and
     batch) of the input.  It may also be a tuple, to control dimensions individually.
@@ -308,12 +308,11 @@ One of two keywords must be given:
 
   - If `scale` is a number, this applies to all but the last two dimensions (channel and
     batch) of the input.  It may also be a tuple, to control dimensions individually.
-  - `mode`: Set to `:nearest`, `:linear`, `:bilinear` or `:trilinear`
+  - `mode`: Set to `:nearest`, `:bilinear` or `:trilinear`
 
 Currently supported upsampling `mode`s and corresponding NNlib's methods are:
 
   - `:nearest` -> `NNlib.upsample_nearest`
-  - `:linear` -> `NNlib.upsample_linear`
   - `:bilinear` -> `NNlib.upsample_bilinear`
   - `:trilinear` -> `NNlib.upsample_trilinear`
 
@@ -323,12 +322,12 @@ Currently supported upsampling `mode`s and corresponding NNlib's methods are:
     function
     
       + As a rule of thumb, `:nearest` should work with arrays of arbitrary dimensions
-      + `:linear` works with 3D Arrays, `:bilinear` works with 4D Arrays, and `:trilinear`
-        works with 5D Arrays
+      + `:bilinear` works with 4D Arrays
+      + `:trilinear` works with 5D Arrays
 
 ## Returns
 
-  - Upsampled Input of size `size` or of size `(I_1 × scale[1], ..., I_N × scale[N], C, N)`
+  - Upsampled Input of size `size` or of size `(I_1 x scale[1], ..., I_N x scale[N], C, N)`
   - Empty `NamedTuple()`
 """
 struct Upsample{mode, S, T} <: AbstractExplicitLayer
@@ -339,7 +338,7 @@ end
 function Upsample(mode::Symbol=:nearest; scale=nothing, size=nothing)
     mode in [:nearest, :bilinear, :trilinear] ||
         throw(ArgumentError("mode=:$mode is not supported."))
-    if !(isnothing(scale) ⊻ isnothing(size))
+    if !xor(isnothing(scale), isnothing(size))
         throw(ArgumentError("Either scale or size should be specified (but not both)."))
     end
     return Upsample{mode, typeof(scale), typeof(size)}(scale, size)
@@ -350,9 +349,8 @@ Upsample(scale, mode::Symbol=:nearest) = Upsample(mode; scale)
 function (m::Upsample{:nearest})(x::AbstractArray, ps, st::NamedTuple)
     return NNlib.upsample_nearest(x, m.scale), st
 end
-function (m::Upsample{:nearest, Int})(x::AbstractArray{T, N}, ps,
-                                      st::NamedTuple) where {T, N}
-    return NNlib.upsample_nearest(x, ntuple(i -> m.scale, N - 2)), st
+function (m::Upsample{:nearest, Int})(x::AbstractArray, ps, st::NamedTuple)
+    return NNlib.upsample_nearest(x, ntuple(i -> m.scale, ndims(x) - 2)), st
 end
 function (m::Upsample{:nearest, Nothing})(x::AbstractArray, ps, st::NamedTuple)
     return NNlib.upsample_nearest(x; size=m.size), st
@@ -400,7 +398,7 @@ See also [`MaxPool`](@ref), [`AdaptiveMaxPool`](@ref), [`GlobalMeanPool`](@ref)
 struct GlobalMaxPool <: AbstractExplicitLayer end
 
 function (g::GlobalMaxPool)(x, ps, st::NamedTuple)
-    return maxpool(x, PoolDims(x, size(x)[1:(end - 2)])), st
+    return maximum(x; dims=1:(ndims(x) - 2)), st
 end
 
 """
@@ -423,7 +421,7 @@ See also [`MeanPool`](@ref), [`AdaptiveMeanPool`](@ref), [`GlobalMaxPool`](@ref)
 struct GlobalMeanPool <: AbstractExplicitLayer end
 
 function (g::GlobalMeanPool)(x, ps, st::NamedTuple)
-    return meanpool(x, PoolDims(x, size(x)[1:(end - 2)])), st
+    return mean(x; dims=1:(ndims(x) - 2)), st
 end
 
 """