+++ test/layers/conv.jl
@@ -10,48 +10,56 @@ include("../test_utils.jl")
     y = randn(rng, Float32, 20, 20, 3, 2)
 
     layer = AdaptiveMaxPool((5, 5))
+    println(layer)
     ps, st = Lux.setup(rng, layer)
 
     @test layer(x, ps, st)[1] == maxpool(x, PoolDims(x, 2))
     run_JET_tests(layer, x, ps, st)
 
     layer = AdaptiveMeanPool((5, 5))
+    println(layer)
     ps, st = Lux.setup(rng, layer)
 
     @test layer(x, ps, st)[1] == meanpool(x, PoolDims(x, 2))
     run_JET_tests(layer, x, ps, st)
 
     layer = AdaptiveMaxPool((10, 5))
+    println(layer)
     ps, st = Lux.setup(rng, layer)
 
     @test layer(y, ps, st)[1] == maxpool(y, PoolDims(y, (2, 4)))
     run_JET_tests(layer, x, ps, st)
 
     layer = AdaptiveMeanPool((10, 5))
+    println(layer)
     ps, st = Lux.setup(rng, layer)
 
     @test layer(y, ps, st)[1] == meanpool(y, PoolDims(y, (2, 4)))
     run_JET_tests(layer, x, ps, st)
 
     layer = GlobalMaxPool()
+    println(layer)
     ps, st = Lux.setup(rng, layer)
 
     @test size(layer(x, ps, st)[1]) == (1, 1, 3, 2)
     run_JET_tests(layer, x, ps, st)
 
     layer = GlobalMeanPool()
+    println(layer)
     ps, st = Lux.setup(rng, layer)
 
     @test size(layer(x, ps, st)[1]) == (1, 1, 3, 2)
     run_JET_tests(layer, x, ps, st)
 
     layer = MaxPool((2, 2))
+    println(layer)
     ps, st = Lux.setup(rng, layer)
 
     @test layer(x, ps, st)[1] == maxpool(x, PoolDims(x, 2))
     run_JET_tests(layer, x, ps, st)
 
     layer = MeanPool((2, 2))
+    println(layer)
     ps, st = Lux.setup(rng, layer)
 
     @test layer(x, ps, st)[1] == meanpool(x, PoolDims(x, 2))
@@ -63,6 +71,7 @@ include("../test_utils.jl")
         x = ones(Float32, (k .+ 3)..., 1, 1)
 
         layer = ltype(k; pad=Lux.SamePad())
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         @test size(layer(x, ps, st)[1])[1:(end - 2)] == cld.(size(x)[1:(end - 2)], k)
@@ -74,6 +83,7 @@ end
     @testset "Grouped Conv" begin
         x = rand(rng, Float32, 4, 6, 1)
         layer = Conv((3,), 6 => 2; groups=2)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         @test size(ps.weight) == (3, 3, 2)
@@ -84,6 +94,7 @@ end
 
         x = rand(rng, Float32, 4, 4, 6, 1)
         layer = Conv((3, 3), 6 => 2; groups=2)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         @test size(ps.weight) == (3, 3, 3, 2)
@@ -94,6 +105,7 @@ end
 
         x = rand(rng, Float32, 4, 4, 4, 6, 1)
         layer = Conv((3, 3, 3), 6 => 2; groups=2)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         @test size(ps.weight) == (3, 3, 3, 3, 2)
@@ -104,13 +116,16 @@ end
 
         # Test that we cannot ask for non-integer multiplication factors
         layer = Conv((2, 2), 3 => 10; groups=2)
+        println(layer)
         @test_throws AssertionError Lux.setup(rng, layer)
         layer = Conv((2, 2), 2 => 9; groups=2)
+        println(layer)
         @test_throws AssertionError Lux.setup(rng, layer)
     end
 
     @testset "Asymmetric Padding" begin
         layer = Conv((3, 3), 1 => 1, relu; pad=(0, 1, 1, 2))
+        println(layer)
         x = ones(Float32, 28, 28, 1, 1)
         ps, st = Lux.setup(rng, layer)
 
@@ -131,16 +146,21 @@ end
 
     @testset "Variable BitWidth Parameters" begin
         # https://github.com/FluxML/Flux.jl/issues/1421
-        layer = Conv((5, 5), 10 => 20, identity; init_weight=Base.randn)
+        layer = Conv((5, 5), 10 => 20, identity; init_weight=Base.randn,
+                     init_bias=(rng, dims...) -> randn(rng, Float16, dims...))
+        println(layer)
         ps, st = Lux.setup(rng, layer)
-        @test ps.bias isa Array{Float64, 4}
+        @test ps.weight isa Array{Float64, 4}
+        @test ps.bias isa Array{Float16, 4}
     end
 
     @testset "Depthwise Conv" begin
         x = randn(rng, Float32, 4, 4, 3, 2)
 
         layer = Conv((2, 2), 3 => 15; groups=3)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
+        @test Lux.parameterlength(layer) == Lux.parameterlength(ps)
 
         @test size(layer(x, ps, st)[1], 3) == 15
         run_JET_tests(layer, x, ps, st)
@@ -148,6 +168,7 @@ end
                                       atol=1.0f-3, rtol=1.0f-3)
 
         layer = Conv((2, 2), 3 => 9; groups=3)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         @test size(layer(x, ps, st)[1], 3) == 9
@@ -155,8 +176,10 @@ end
         test_gradient_correctness_fdm((x, ps) -> sum(layer(x, ps, st)[1]), x, ps;
                                       atol=1.0f-3, rtol=1.0f-3)
 
-        layer = Conv((2, 2), 3 => 9; groups=3, bias=false)
+        layer = Conv((2, 2), 3 => 9; groups=3, use_bias=false)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
+        @test Lux.parameterlength(layer) == Lux.parameterlength(ps)
 
         @test size(layer(x, ps, st)[1], 3) == 9
         run_JET_tests(layer, x, ps, st)
@@ -165,6 +188,7 @@ end
 
         # Test that we cannot ask for non-integer multiplication factors
         layer = Conv((2, 2), 3 => 10; groups=3)
+        println(layer)
         @test_throws AssertionError Lux.setup(rng, layer)
     end
 
@@ -172,6 +196,7 @@ end
         x = ones(Float32, (k .+ 3)..., 1, 1)
 
         layer = Conv(k, 1 => 1; pad=Lux.SamePad())
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         @test size(layer(x, ps, st)[1]) == size(x)
@@ -180,6 +205,7 @@ end
                                       atol=1.0f-3, rtol=1.0f-3)
 
         layer = Conv(k, 1 => 1; pad=Lux.SamePad(), dilation=k .÷ 2)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         @test size(layer(x, ps, st)[1]) == size(x)
@@ -189,6 +215,7 @@ end
 
         stride = 3
         layer = Conv(k, 1 => 1; pad=Lux.SamePad(), stride=stride)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         @test size(layer(x, ps, st)[1])[1:(end - 2)] == cld.(size(x)[1:(end - 2)], stride)
@@ -202,6 +229,7 @@ end
         x[4, 4, 1, 1] = 1
 
         layer = Conv((3, 3), 1 => 1)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         y = zeros(eltype(ps.weight), 5, 5, 1, 1)
@@ -210,6 +238,7 @@ end
         run_JET_tests(layer, x, ps, st)
 
         layer = Conv((3, 1), 1 => 1)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         y = zeros(eltype(ps.weight), 5, 7, 1, 1)
@@ -218,6 +247,7 @@ end
         run_JET_tests(layer, x, ps, st)
 
         layer = Conv((1, 3), 1 => 1)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         y = zeros(eltype(ps.weight), 7, 5, 1, 1)
@@ -226,11 +256,91 @@ end
         run_JET_tests(layer, x, ps, st)
 
         layer = Conv((1, 3), 1 => 1; init_weight=Lux.glorot_normal)
+        println(layer)
         ps, st = Lux.setup(rng, layer)
 
         y = zeros(eltype(ps.weight), 7, 5, 1, 1)
         y[4, 2:(end - 1), 1, 1] = ps.weight
         @test y ≈ layer(x, ps, st)[1]
         run_JET_tests(layer, x, ps, st)
+    end
+
+    # Deprecated Functionality (Remove in v0.5)
+    @testset "Deprecations" begin
+        @test_deprecated layer = Conv((3, 3), 1 => 1; bias=false)
+        ps, st = Lux.setup(rng, layer)
+        @test !hasproperty(ps, :bias)
+
+        @test_throws ArgumentError layer=Conv((3, 3), 1 => 1; bias=false, use_bias=false)
+    end
+end
+
+@testset "Upsample" begin
+    @testset "Construction" begin
+        @test_nowarn Upsample(:nearest; scale=2)
+        @test_nowarn Upsample(:nearest; size=(64, 64))
+        @test_nowarn Upsample(:bilinear; scale=2)
+        @test_nowarn Upsample(:bilinear; size=(64, 64))
+        @test_nowarn Upsample(:trilinear; scale=2)
+        @test_nowarn Upsample(:trilinear; size=(64, 64))
+
+        @test_throws ArgumentError Upsample(:linear; scale=2)
+        @test_throws ArgumentError Upsample(:nearest; scale=2, size=(64, 64))
+        @test_throws ArgumentError Upsample(:nearest)
+
+        @test_nowarn Upsample(2)
+        @test_nowarn Upsample(2, :nearest)
+    end
+
+    @testset "Size Correctness" begin
+        # NNlib is checking algorithmic correctness. So we should just verify correct
+        # function call
+        modes = (:nearest, :bilinear, :trilinear)
+        sizes = (nothing, (64, 64), (64, 32))
+        scales = (nothing, 2, (2, 1))
+
+        for mode in modes, xsize in sizes, scale in scales
+            if !xor(isnothing(xsize), isnothing(scale))
+                continue
+            end
+            layer = Upsample(mode; size=xsize, scale=scale)
+            println(layer)
+            ps, st = Lux.setup(rng, layer)
+            x = zeros((32, 32, 3, 4))
+
+            run_JET_tests(layer, x, ps, st)
+
+            y, _ = layer(x, ps, st)
+            if isnothing(scale)
+                @test size(y)[1:2] == xsize
+            else
+                @test size(y)[1:2] == size(x)[1:2] .* scale
+            end
+            @test size(y)[3:4] == size(x)[3:4]
+        end
+
+        sizes = (nothing, (64, 64, 64), (64, 32, 128))
+        scales = (nothing, 2, (2, 1, 1), (2, 2, 1))
+
+        for mode in modes, xsize in sizes, scale in scales
+            if !xor(isnothing(xsize), isnothing(scale))
+                continue
+            end
+            layer = Upsample(mode; size=xsize, scale=scale)
+            println(layer)
+            ps, st = Lux.setup(rng, layer)
+            x = zeros((32, 32, 32, 3, 4))
+
+            run_JET_tests(layer, x, ps, st)
+
+            y, _ = layer(x, ps, st)
+
+            if isnothing(scale)
+                @test size(y)[1:3] == xsize
+            else
+                @test size(y)[1:3] == size(x)[1:3] .* scale
+            end
+            @test size(y)[4:5] == size(x)[4:5]
+        end
     end
 end